requirements.txt
arxiv==2.1.3
openai>=1.0.0
python-dotenv>=1.0.0
aiohttp>=3.8.0
matplotlib>=3.5.0
numpy>=1.21.0

